- Added a helper method in runner.py to avoid direct access when populate the list with functions. - Wrote a filler-template for the context-menu - Re-did some functions for clarity
